lua脚本解决超卖。同时引入redis作为存储

+++ ms-seckill/src/main/java/com/lwq/seckill/mapper/VoucherOrdersMapper.java
@@ -23,6 +23,6 @@ public interface VoucherOrdersMapper {
             " status, fk_seckill_id, order_type, create_date, update_date,  is_valid)" +
             " values (#{orderNo}, #{fkVoucherId}, #{fkDinerId}, #{status}, #{fkSeckillId}, " +
             " #{orderType}, now(), now(), 1)")
-    int save(VoucherOrders voucherOrders);
+    long save(VoucherOrders voucherOrders);
 
 }
